# docs/swagger.yaml
definitions:
  input.UserInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
host: backend-annisa-production.up.railway.app
info:
  contact: {}
  description: Sweager service API in Go using Gin framework
  title: Sweager Service API
paths:
  /api/blog:
    get:
      consumes:
      - application/json
      description: Retrieve all blogs with pagination
      parameters:
      - description: Page number for pagination (default is 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default is 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Get all blogs
      tags:
      - Blog
    post:
      consumes:
      - application/json
      description: Create New blog
      parameters:
      - description: File gambar
        in: formData
        name: file
        required: true
        type: file
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create New blog
      tags:
      - blog
  /api/blog/{slug}:
    delete:
      consumes:
      - application/json
      description: Delete a blog by its slug
      parameters:
      - description: Slug of the blog to be deleted
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a blog by slug
      tags:
      - Blog
    get:
      consumes:
      - application/json
      description: Retrieve a single blog using its slug
      parameters:
      - description: Slug of the blog
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Get a single blog by slug
      tags:
      - Blog
    put:
      consumes:
      - application/json
      description: Update blog by slug
      parameters:
      - description: slug blog
        in: path
        name: slug
        required: true
        type: string
      - description: File gambar
        in: formData
        name: file
        required: true
        type: file
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update blog by slug
      tags:
      - blog
  /api/treatment:
    get:
      consumes:
      - application/json
      description: Get All treatment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Get All treatment
      tags:
      - treatment
    post:
      consumes:
      - application/json
      description: Create New treatment
      parameters:
      - description: TreatmentName
        in: formData
        name: treatment_name
        required: true
        type: string
      - description: Price
        in: formData
        name: price
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create New treatment
      tags:
      - Treatment
  /api/treatment/{slug}:
    delete:
      consumes:
      - application/json
      description: Delete a treatment by its slug
      parameters:
      - description: Slug of the treatment to be deleted
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a treatment by slug
      tags:
      - treatment
    put:
      consumes:
      - application/json
      description: Update treatment by slug
      parameters:
      - description: slug treatment
        in: path
        name: slug
        required: true
        type: string
      - description: TreatmentName
        in: formData
        name: treatment_name
        required: true
        type: string
      - description: Price
        in: formData
        name: price
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update treatment by slug
      tags:
      - Treatment
  /api/treatments/{slug}:
    get:
      consumes:
      - application/json
      description: Retrieve a single treatment using its slug
      parameters:
      - description: Slug of the treatment
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Get a single treatment by slug
      tags:
      - treatment
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Log in an existing user using email and password
      parameters:
      - description: User login details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/input.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - Users
  /api/user/register:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided information
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/input.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register new user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"